=2019.8.28

+++ auto-commit/index.js
@@ -1,7 +1,23 @@
 const cmd = require('./cmd');
-; (async () => {
+const file = require('./file');
+//昨天的时间
+const today = new Date();
+let day = 1;
+const commit = async () => {
+    today.setTime(today.getTime() - day * 24 * 60 * 60 * 1000);
+    let commitTime = `${today.getFullYear()}.${today.getMonth() + 1}.${today.getDate()}`;
+    console.log(commitTime);
+    await file();
     await cmd('git status');
     await cmd('git add .');
-    await cmd('git commit --amend --no-edit --date="Sun, 25 Dec 2016 19:42:09 +0800"');
-    await cmd('git push origin master --force');
-})();
+    // await cmd(`git commit --amend --no-edit --date="Sun, 25 Dec 2016 19:42:09 +0800"`);
+    // await cmd(`git commit --amend --no-edit --date="${Date()}"`);
+    await cmd(`git commit -m="${commitTime}" --amend --no-edit --date="${commitTime}"`);
+    if (day > 10) {
+        await cmd('git push origin master --force');
+    } else {
+        day += 1;
+        commit();
+    }
+}
+commit();
